Some changes in secondary ice processes: the contributions are now calculated before applied to the  ice size distributions. Added the Phillips et al parameterization for drop fragmentation. Crashing issues  realted to secondary ice *seem* to have been resolved, but still testing.

+++ src/src_LES/mpi_interface.f90
@@ -956,7 +956,7 @@ CONTAINS
    IMPLICIT NONE
 
    REAL, INTENT(inout) :: sendbuff
-   REAL, INTENT(in) :: rootid
+   INTEGER, INTENT(in) :: rootid
    INTEGER :: ierror
 
    CALL MPI_BCAST(sendbuff,1,MY_REAL,rootid,MPI_COMM_WORLD,ierror)
